+++ src/OpenXR/Silk.NET.OpenXR/Enums/DebugUtilsMessageTypeFlagsEXT.gen.cs
@@ -11,7 +11,7 @@ namespace Silk.NET.OpenXR
 {
     [Flags()]
     [NativeName("Name", "XrDebugUtilsMessageTypeFlagsEXT")]
-    public enum DebugUtilsMessageTypeFlagsEXT : int
+    public enum DebugUtilsMessageTypeFlagsEXT : long
     {
         [NativeName("Name", "XR_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT")]
         DebugUtilsMessageTypeGeneralBitExt = 1,
