+++ src/OpenXR/Silk.NET.OpenXR/Enums/OverlayMainSessionFlagsEXTX.gen.cs
@@ -11,7 +11,7 @@ namespace Silk.NET.OpenXR
 {
     [Flags()]
     [NativeName("Name", "XrOverlayMainSessionFlagsEXTX")]
-    public enum OverlayMainSessionFlagsEXTX : int
+    public enum OverlayMainSessionFlagsEXTX : long
     {
         [NativeName("Name", "XR_OVERLAY_MAIN_SESSION_ENABLED_COMPOSITION_LAYER_INFO_DEPTH_BIT_EXTX")]
         OverlayMainSessionEnabledCompositionLayerInfoDepthBitExtx = 1,
